Seperated react src files
