+++ packages/inpage-nav/src/react/inpage-nav.js
@@ -18,66 +18,6 @@ import PropTypes from 'prop-types';
 
 
 /**
- * The section component for the AUinpageNavLinks component
- *
- * @param  {object} link             - The link of this section
- * @param  {object} title            - The title of the section
- * @param  {object} li               - An additional object to be spread into the wrapping element, optional
- * @param  {object} attributeOptions - Any other attribute options
- */
-export const AUinpageNavLinksItem = ({ link, title, li = {}, ...attributeOptions }) => (
-	<li { ...li }>
-		<a href={ `#${ link }` } { ...attributeOptions }>{ title }</a>
-	</li>
-);
-
-AUinpageNavLinksItem.propTypes = {
-	link: PropTypes.string.isRequired,
-	title: PropTypes.string.isRequired,
-	li: PropTypes.object,
-};
-
-
-/**
- * The inpage-nav component
- *
- * @param  {boolean} dark             - Add the dark variation class, optional
- * @param  {string}  title            - The title of the content link block, default: Contents
- * @param  {array}   sections         - An array of objects of all sections, sample: { link: '', title: '', onClick: () }
- * @param  {string}  className        - An additional class, optional
- * @param  {string}  ariaLabel        - The aria-label attribute, optional
- * @param  {object}  attributeOptions - Any other attribute options
- */
-export const AUinpageNavLinks = ({ dark, title, sections, ariaLabel, className = '', ...attributeOptions }) => (
-	<nav className={ `au-inpage-nav-links ${ className }${ dark ? ' au-inpage-nav-links--dark' : '' }` } ariaLabel={ ariaLabel } { ...attributeOptions }>
-		<h2 className="au-inpage-nav-links__heading au-display-sm">{ title }</h2>
-
-		<ul className="au-link-list">
-			{ sections.map( ( section, i ) => <AUinpageNavLinksItem key={ i } { ...section } /> ) }
-		</ul>
-	</nav>
-);
-
-AUinpageNavLinks.propTypes = {
-	dark: PropTypes.bool,
-	title: PropTypes.string.isRequired,
-	sections: PropTypes.arrayOf(
-		PropTypes.shape({
-			link: PropTypes.string.isRequired,
-			title: PropTypes.string.isRequired,
-			li: PropTypes.object,
-		})
-	).isRequired,
-	className: PropTypes.string,
-};
-
-AUinpageNavLinks.defaultProps = {
-	title: 'Contents',
-	ariaLabel: 'in page navigation',
-};
-
-
-/**
  * The inpage-nav component
  *
  * @param  {string}   title            - The title of the section
